infrastructure/pyproject.toml
[tool.poetry]
name = "infrastructure"
version = "0.1.0"
description = ""
authors = ["gersonUb <gersoncabj@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app_infrastructure", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.46"
alembic = "^1.18.1"
asyncpg = "^0.31.0"
langchain = "^1.2.7"
langchain-community = "^0.4.1"
langchain-chroma = "^1.1.0"
langchain-huggingface = "^1.2.0"
ollama = "^0.6.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
email-validator = "^2.3.0"
app-domain = {path = "../domain", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"